Time: 7 ms (95.52%) | Memory: 64.1 MB (66.83%) - LeetSync

+++ 969-number-of-recent-calls/number-of-recent-calls.cpp
@@ -9,7 +9,7 @@ public:
     int ping(int t) {
         q.push(t);
         while(!q.empty() && q.front()<t-3000)q.pop();
-        return static_cast<int>(q.size());
+        return (q.size());
     }
 };
